Module wld::models

Structs

AnglerQuestStatus
A bucket for a few angler-related things.

Backgrounds

Block
Represents a block. A Tile optionally "has" a Block.

BlockCtx
Contains some metadata from the Tile instance to which a Block belongs.

BossesSlain

BossesSlain2

Bounds
Represents the size of a Terraria map, as a rectangle.

Chest
Represents a chest.

Chests
Represents all Chests in a Terraria world.

Clouds

EventsCompleted

ExtendedTileAttributes

Footer

GeneratorInfo
Information about the world seed.

HardmodeOre
Represents one of three types of ore available in hardmode.

Header
World file header.

House
Represents NPC housing.

HouseVec
A variable-length list of Houses.

InvasionStatus

ItemStack
Represents a stack of items in a Chest.

ItemStackVec
A list of ItemStacks, which make up the contents of a Chest.

Liquid

LogicSensor

Mob
Represents a "mob"; a non-friendly NPC entity.

MobKills
A list of kill counts corresponding to mobs.

MobVec
List of Mobs; fancy wrapper around a Vec.

NPC
Represents a friendly NPC.

NPCVec

Offsets
Byte offsets in a .wld file corresponding to certain data points.

OldOnesArmyStatus
Represents the unlocked tiers of the Old One's Army event.

PartyStatus
Various bits of party-related information.

PartyingNPCs
A list of partying NPCs.

PillarStatus

Position
Represents a position (or point) in terms of cartesian coordinates.

PressurePlates
Represents all PressurePlates in the game world.

Properties

QuadrantStyle
Represents styles which differ by "quandrant" of the map (by X-axis).

RainStatus

RunLength

SandstormStatus
Bucket for info about sandstorm events.

SavedNPCs

ShadowOrbStatus

Sign

Signs

Status

TBitVec
A fixed-length vector of bits; a fancy wrapper around BitVec.

TString
Represents a string as per the data format; a fancy wrapper around String.

TUuid
A UUID.

Tile

TileAttributes

TileCtx

TileEntities

TileEntity

TileHeader

TileMatrix

TileVec

TownManager

Uuid
A Universally Unique Identifier (UUID).

VariableTBitVec
A variable-length vector of bits; a fancy wrapper around BitVec.

Wall

Wiring
Represents wires on a Tile. A Tile can have all of these, none of these, or any in between. For more information, see Wire on the Official Terraria Wiki.

World
Top-level representation of a Terraria world file (.wld).

WorldCtx

WorldStyle
A bucket in which some style information lives.

Type Definitions

HellStyle
Represents the appearance of the hell biome's background.

MoonStyle
Represents the appearance of the moon.

PressurePlate
Represents a pressure plate.

UndergroundJungleStyle
Represents the appearance of the underground jungle biome's background.

UndergroundSnowStyle
Represents the appearance of the underground snow biome's background.